+++ src/main/kotlin/com/github/sleepypanda/feesh/features/rendering/HideTadgangNametags.kt
@@ -22,14 +22,14 @@ object HideTadgangNametags {
     @JvmStatic
     fun shouldHide(entityId: Int): Boolean {
         if (!WorldRendering.hideTadgangNametags) return false
-        if (!WorldUtils.isInSkyblock() || WorldUtils.getWorldName() != WorldUtils.GALATEA) return false
+        if (!WorldUtils.isInSkyblock() || !isInGalatea()) return false
         return entityId in hiddenNametagIds
     }
 
     private fun onArmorStandCustomNameChanged(event: ArmorStandCustomNameChangedEvent) {
         if (!event.isFirstLoaded) return
         if (!WorldRendering.hideTadgangNametags) return
-        if (!WorldUtils.isInSkyblock() || WorldUtils.getWorldName() != WorldUtils.GALATEA) return
+        if (!WorldUtils.isInSkyblock() || !isInGalatea()) return
 
         CommonUtils.runWithCatching("Failed to check Tadgang nametag") {
             val name = event.customName.unformatted
@@ -45,5 +45,9 @@ object HideTadgangNametags {
 
     private fun onWorldChanged(@Suppress("UNUSED_PARAMETER") event: WorldChangedEvent) {
         hiddenNametagIds.clear()
+    }
+
+    private fun isInGalatea(): Boolean {
+        return WorldUtils.getWorldName() == WorldUtils.GALATEA || WorldUtils.getWorldName() == WorldUtils.MOONGLADE_MARSH
     }
 }