+++ src/main/kotlin/com/github/sleepypanda/feesh/features/sounds/MuteJadeDragonSound.kt
@@ -8,11 +8,15 @@ object MuteJadeDragonSound {
     @JvmStatic
     fun shouldCancel(soundId: Identifier?): Boolean {
         if (!WorldRendering.muteJadeDragon) return false
-        if (!WorldUtils.isInSkyblock() || WorldUtils.getWorldName() != WorldUtils.GALATEA || 
+        if (!WorldUtils.isInSkyblock() || !isInGalatea() || 
             (WorldUtils.getZoneName() != WorldUtils.DRAGON_LAIR && WorldUtils.getZoneName() != WorldUtils.MURKWATER_DEPTHS)
         ) return false
         if (soundId == null || soundId.namespace != "minecraft" || !soundId.path.startsWith("entity.ender_dragon.")) return false
 
         return true
+    }
+
+    private fun isInGalatea(): Boolean {
+        return WorldUtils.getWorldName() == WorldUtils.GALATEA || WorldUtils.getWorldName() == WorldUtils.MOONGLADE_MARSH
     }
 }