+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/models/BulkResettableTrackerTypes.kt
@@ -5,7 +5,7 @@ import com.github.sleepypanda.feesh.features.overlays.base.IResettableViewModeTr
 import com.github.sleepypanda.feesh.features.overlays.BayouTracker
 import com.github.sleepypanda.feesh.features.overlays.CrimsonIsleTracker
 import com.github.sleepypanda.feesh.features.overlays.FishingProfitTracker
-import com.github.sleepypanda.feesh.features.overlays.GalateaWaterTracker
+import com.github.sleepypanda.feesh.features.overlays.MoongladeMarshWaterTracker
 import com.github.sleepypanda.feesh.features.overlays.JerryWorkshopTracker
 import com.github.sleepypanda.feesh.features.overlays.LotusAtollTracker
 import com.github.sleepypanda.feesh.features.overlays.TorrhusCanyonTracker
@@ -29,7 +29,7 @@ enum class BulkResettableTrackerTypes(
     BAYOU_TRACKER("Bayou tracker", resettableTracker = BayouTracker),
     WATER_HOTSPOTS_TRACKER("Water Hotspots tracker", resettableTracker = WaterHotspotsTracker),
     CRIMSON_ISLE_TRACKER("Crimson Isle tracker", resettableTracker = CrimsonIsleTracker),
-    GALATEA_WATER_TRACKER("Galatea water tracker", resettableTracker = GalateaWaterTracker),
+    GALATEA_WATER_TRACKER("Moonglade Marsh water tracker", resettableTracker = MoongladeMarshWaterTracker),
     TORRHUS_CANYON_TRACKER("Torrhus Canyon tracker", resettableTracker = TorrhusCanyonTracker),
     LOTUS_ATOLL_TRACKER("Lotus Atoll tracker", resettableTracker = LotusAtollTracker),
     TREASURE_FISHING_TRACKER("Treasure fishing tracker", hasSessionMode = true, resettableViewModeTracker = TreasureFishingTracker),